requirements.txt
paho-mqtt==1.5.1
pybluez==0.23